+++ src/js/util/options.js
@@ -1,4 +1,4 @@
-import {assign, hasOwn, includes, isArray, isFunction, isUndefined, startsWith} from './lang';
+import {assign, hasOwn, includes, isArray, isFunction, isUndefined, sortBy, startsWith} from './lang';
 
 const strats = {};
 
@@ -27,7 +27,7 @@ strats.destroy = function (parentVal, childVal) {
 
 // update strategy
 strats.update = function (parentVal, childVal) {
-    return strats.args(parentVal, isFunction(childVal) ? {read: childVal} : childVal);
+    return sortBy(strats.args(parentVal, isFunction(childVal) ? {read: childVal} : childVal), 'order');
 };
 
 // property strategy