+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Fix `UIkit.mixin()` 
 - Fix issue with list bullet in combination with multi column layouts
 - Allow for special characters in URI hash in Scroll Component
+- Height Match also sets a `height` in IE
 
 ## 3.0.0 rc 16 (September 12, 2018)